+++ dca/tl_news_category.php
@@ -132,7 +132,7 @@ $GLOBALS['TL_DCA']['tl_news_category'] = array
     // Palettes
     'palettes' => array
     (
-        'default'                     => '{title_legend},title,frontendTitle,alias;{redirect_legend:hide},jumpTo;{publish_legend},published'
+        'default'                     => '{title_legend},title,alias,frontendTitle,cssClass;{redirect_legend:hide},jumpTo;{publish_legend},published'
     ),
 
     // Fields
@@ -160,7 +160,7 @@ $GLOBALS['TL_DCA']['tl_news_category'] = array
             'exclude'                 => true,
             'search'                  => true,
             'inputType'               => 'text',
-            'eval'                    => array('mandatory'=>true, 'maxlength'=>255, 'tl_class'=>'clr'),
+            'eval'                    => array('mandatory'=>true, 'maxlength'=>255, 'tl_class'=>'w50'),
             'sql'                     => "varchar(255) NOT NULL default ''"
         ),
         'frontendTitle' => array
@@ -184,6 +184,14 @@ $GLOBALS['TL_DCA']['tl_news_category'] = array
                 array('tl_news_category', 'generateAlias')
             ),
             'sql'                     => "varbinary(128) NOT NULL default ''"
+        ),
+        'cssClass' => array
+        (
+            'label'                   => &$GLOBALS['TL_LANG']['tl_news_category']['cssClass'],
+            'exclude'                 => true,
+            'inputType'               => 'text',
+            'eval'                    => array('maxlength'=>128, 'tl_class'=>'w50'),
+            'sql'                     => "varchar(128) NOT NULL default ''",
         ),
         'jumpTo' => array
         (