+++ languages/pl/tl_news_category.php
@@ -18,6 +18,7 @@
 $GLOBALS['TL_LANG']['tl_news_category']['title']         = array('Tytuł', 'Wprowadź tytuł kategorii.');
 $GLOBALS['TL_LANG']['tl_news_category']['frontendTitle'] = array('Tytuł na stronie', 'Tutaj możesz wprowadzić tytuł kategorii, który będzie widoczny na stronie.');
 $GLOBALS['TL_LANG']['tl_news_category']['alias']         = array('Alias kategorii', 'Alias kategorii jest unikalnym odwołaniem do kategorii, do którego można się odwołać zamiast numerycznego ID.');
+$GLOBALS['TL_LANG']['tl_news_category']['cssClass']      = array('Klasa CSS', 'Tutaj możesz wprowadzić klasę CSS, która będzie dodana we front endzie.');
 $GLOBALS['TL_LANG']['tl_news_category']['jumpTo']        = array('Strona przekierowania', 'Tutaj możesz wybrać stronę, na którą zostanie przeniesiony odwiedzający po kliknięciu na link kategorii w szablonie aktualności.');
 $GLOBALS['TL_LANG']['tl_news_category']['published']     = array('Opublikuj kategorię', 'Opublikuj kategorię aktualności na stronie.');